==> 01_Web Scraping Rayados.ipynb <==
# Web Scrapping Datos Rayados Wikipedia

Se importan Librerias

from bs4 import BeautifulSoup
import pandas as pd
import requests

Carga Y Validacion URL

# URL donde vienen las estadisticas de Rayados, tanto
url = "https://es.wikipedia.org/wiki/Anexo:Estad%C3%ADsticas_del_Club_de_F%C3%BAtbol_Monterrey"
page = requests.get(url)
page

# Validacion Codigo Pagina Web 
soup = BeautifulSoup(page.text, "html")
print(soup.prettify())

## Tabla Liga

### Extraer Informacion de la Pagina para crear el Primer DataFrame

# Traer Informacion de la Tabla con informacion de las Ligas, es la Sexta Tabla de toda la Pagina.
tabla_torneo = soup.find_all('table')[6:7]
#tabla_torneo

#Traer Encabezados de la Tabla Liga
for enc_liga in tabla_torneo:
    encabezado_liga = enc_liga.find_all('th')
columnas_liga = [columna.text.strip() for columna in encabezado_liga ]

### Creacion y Validacion Data Frame

Llenar el DataFrame primero con la informacion de las Columnas

df_liga = pd.DataFrame(columns = columnas_liga)
df_liga

Llenado del DataFrame con la informacion de la Tabla

#Traer Informacion de la Tabla al DataFrame
for tabla in tabla_torneo:
    for row in tabla.find_all('tr')[1:]:
        values = []
        for columns in row.find_all('td'):
            values.append(columns.text.strip())
            #print(values)
            #print(len(values))
            if len(values) == len(columnas_liga):
                df_liga.loc[len(df_liga)] = values

# Validacion de la Informmacion
df_liga

### Guardado de la Informacion en un CSV

# Guardar el Data Frame en un CSV
df_liga.to_csv(r'C:\Users\Jorge\Documents\Portafolio V2\Python\Python_03_Rayados\Estadisticas_Rayados_liga.csv',index = False)

# Tabla Liguilla

## Extraccion de la Informacion

#Extraer Codigo de la Tabla con informacion de las Liguillas.
table_liguilla = soup.find_all('table')[8:9]
print(table_liguilla)

# Consegiuir la informacion de los Encabezados
for enc in table_liguilla:
    enc_liguilla = enc.find_all('th')
columnas_liguilla = [punt.text.strip() for punt in enc_liguilla ]

### Creacion del DataFrame

Llenar el DataFrame primero con la informacion de las Columnas

df_liguilla = pd.DataFrame(columns = columnas_liguilla)
df_liguilla

Llenar el DataFrame primero con la informacion de la Tabla

#Extrae la informacion de la Tabla
for table_l in table_liguilla:
    for row in table_l.find_all('tr')[1:]:
        values_l = []
        for col in row.find_all('td'):
            values_l.append(col.text.strip())
            if len(values_l) == len(columnas_liguilla):
                df_liguilla.loc[len(df_liguilla)] = values_l

# Validar Informacion del Data Frame
df_liguilla

### Exportacion del DataFrame a aun CSV

#Exportar Archivo a un CSV
df_liguilla.to_csv(r'C:\Users\Jorge\Documents\Portafolio V2\Python\Python_03_Rayados\Estadisticas_Rayados_liguilla.csv',index = False)
