==> 02_Rayados_Liga_Cleaning.ipynb <==
#LIBRERIAS
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

## Limpieza DF Liga

# Data Frame Liga
df_liga = pd.read_csv(r'C:\Users\Jorge\Documents\Portafolio V2\Python\Python_03_Rayados\Estadisticas_Rayados_liga.csv')

# Informacion de los Tipos de Datos del DataFrame
df_liga.info()

#Definir el Indice del DF al campo #, el cual
df_liga.set_index('#')

### Limpieza de la Columna Torneo y columna Tipo_Torneo
- Algunos Datos traian la letra 'I' al final,
- Se agrego una nueva columna con el nombre de la Columna a Tipo_Torneo, para poder distinguir los que son Clausura / Apertura / Ligas (Primeros torneos)
- Algunos Torneos Tuvieron nombres diferentes, se estandarizaron los nombres para poder agruparse despues.

# Se elimino la letra I al final del los torneos
df_liga['Torneo'] = df_liga['Torneo'].str.rstrip('I')

# Cambio de nombre a la columna
df_liga['Tipo_Torneo'] = df_liga['Torneo']

# Correccion de Nombres
df_liga['Tipo_Torneo'] = df_liga['Tipo_Torneo'].str.replace('Invierno','Apertura')
df_liga['Tipo_Torneo'] = df_liga['Tipo_Torneo'].str.replace('Verano','Clausura')
df_liga['Tipo_Torneo'] = df_liga['Tipo_Torneo'].str.replace('Bicentenario','Clausura')

# Correcion de Nombre Especifica
df_liga['Tipo_Torneo'].loc[40] = 'Apertura'

# Los torneos primeros 39 torneos se agruparan como Ligas
df_liga['Tipo_Torneo'].loc[0:40] = 'Liga'

#Se añade otra columna para poder separar los años de los torneos Clausura / Apertura y dejar la columa Libre
df_liga[['Tipo_Torneo','Basura']] =df_liga['Tipo_Torneo'].str.split(' ', n = 2, expand = True)

#Se Elimina la columna resultante
df_liga = df_liga.drop(columns= 'Basura')

#Se creo una columna nueva llamada Año, para los torneos que califican como Liga, se extrajeron solo los primeros 3 digitos,
for x in df_liga.index:
    #print(df_liga.loc[x, 'Tipo_Torneo'])    
    if df_liga.loc[x, 'Tipo_Torneo'] == 'Liga':
        df_liga.loc[x,"Año"] = df_liga.loc[x, 'Torneo'][:4].strip()
    

#Mismo proceso para los torneos Liga con nombres extraños
df_liga["Año"] = df_liga["Año"].str.replace("Prod",'1985')
df_liga.loc[12,"Año"] = 1970
df_liga.loc[29,"Año"] = 1986
df_liga.loc[40,"Año"] = 1996

#Se quitan los valores NAN por '' para poder cambiarles el valor
df_liga = df_liga.fillna('')

# Se recorre el DF y los campos '' de la columna Año se poblaran por los ultimos 4 digitos correspondientes al año de los torneos
for x in df_liga.index:
    if df_liga.loc[x,"Año"] == '':
        df_liga.loc[x,"Año"] = df_liga.loc[x,"Torneo"][-4:].strip()

## Otras Columnas

#Se eliminA el caracter '°' de la Columna POS 
df_liga['POS'] = df_liga['POS'].str.replace('°','')

#Se cambio el tipo de dato de la columna Año
df_liga['Año'] = df_liga['Año'].astype('int64')

# Calcule la efecitividad de los puntos conseguidos, tomando en cuenta que los torneos antes de 1994 daban 2 puntos y posterior daban 3.
# Hice esto para comprobar que la columna de EFE en realidad calculara la ifnormacion de forma correcta
for x in df_liga.index:
    if df_liga.loc[x,'Año'] <= 1994:
        total_puntos = df_liga.loc[x,"PJ"] * 2
        df_liga.loc[x,"Efectividad"] = efectividad = (df_liga.loc[x,'PTS'] / total_puntos) *100
    else:
        total_puntos = df_liga.loc[x,"PJ"] * 3
        df_liga.loc[x,"Efectividad"] = efectividad = (df_liga.loc[x,'PTS'] / total_puntos) *100
    print(total_puntos, efectividad)

# Añadi una nueva columna para poder agrupar mas adelante los torneos por Decadas.
for x in df_liga.index:
    for y in range(1940,2031,10):
        if df_liga.loc[x,'Año'] >= y and df_liga.loc[x,'Año'] <= y+9:
            #Decadas = print'Decada', y)
            df_liga.loc[x,"Decadas"] = 'Decada ' + str(y)

df_liga

#Se exporta el DF
df_liga.to_csv(r'C:\Users\Jorge\Documents\Portafolio V2\Python\Python_03_Rayados\Estadisticas_Rayados_liga_clean.csv',index = False)
