==> 03_Rayados_Liguilla_Cleaning.ipynb <==
Limpieza Archivo 2

#LIBRERIAS
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

#Se importa el documento con informacion de las Liguillas
df_liguilla = pd.read_csv(r'C:\Users\Jorge\Documents\Portafolio V2\Python\Python_03_Rayados\Estadisticas_Rayados_liguilla.csv')

# Mostrar informacion de los tipos de columna
df_liguilla.info()

## Nueva Columna

# Añadi una nueva columna llamada Campeon, para poder marcar unicamente los torneos donde haya sido campeon el equipo
df_liguilla['Campeon'] = df_liguilla['Puestoalcanzado']

for x in df_liguilla.index:
    if df_liguilla.loc[x,'Campeon'] != "Final (Campeón)":
        #campeon = 'Y'
        print(df_liguilla.loc[x,'Campeon'])
        df_liguilla.loc[x,'Campeon'] = 'N'
    else:
        df_liguilla.loc[x,'Campeon'] = 'Y'
    #else:
     #   df_liguilla[x,"Campeon"] = 'N'

df_liguilla

# Limpieza columna Puesto Alcanzado, Quita las palabras (Campeon) y SubCampeon, paara dejar unicamente Final
for x in df_liguilla.index:
    if df_liguilla.loc[x, "Puestoalcanzado"][0:5] == 'Final':
        df_liguilla.loc[x, "Puestoalcanzado"] = df_liguilla.loc[x, "Puestoalcanzado"][0:5]

#Quite la columna de efectividad ya que personalmente para la liguilla no la considero importante.
df_liguilla.drop(columns = 'EFE')

#Exporte el DataFrame en un CSV
df_liguilla.to_csv(r'C:\Users\Jorge\Documents\Portafolio V2\Python\Python_03_Rayados\Estadisticas_Rayados_liguilla_clean.csv',index = False)

#Importe el Dataframe Limpio de las Ligas
df_liga = pd.read_csv(r'C:\Users\Jorge\Documents\Portafolio V2\Python\Python_03_Rayados\Estadisticas_Rayados_liga_clean.csv')

# Y junte ambos DataFrame en uno solo para poder trabajarlos, junte el dataframe Liguilla al de Liga mediante la columna torneo
# en un Left ya que no todos los torneos pasaron a Liguilla
df_completo = df_liga.merge(df_liguilla[['Torneo','PJ','PG','PE','PP','GF','GC','DIF','Puestoalcanzado','Campeon']], how = 'left', on = 'Torneo',suffixes =['_FR','_L'])

# Rellene los campos de estaidisticas de la Liguilla con valores 0 (Ya que no la jugaron)
df_completo[['PJ_L','PG_L','PE_L','PP_L','GF_L','GC_L','DIF_L']] = df_completo[['PJ_L','PG_L','PE_L','PP_L','GF_L','GC_L','DIF_L']].fillna('0')

#Los campos vacios de la columna puesto alcanzado pasan con No Califico en luigar de NAN y la columna Campeon con N
df_completo['Puestoalcanzado'] = df_completo['Puestoalcanzado'].fillna('No Califico')
df_completo['Campeon'] = df_completo['Campeon'].fillna('N')

# Se define el nuevo indice
df_completo = df_completo.set_index('#')

#Quite la columna efectividad para quedarme con la columna que calcule
df_completo = df_completo.drop(columns='EFE')

# Cambie el tipo de dato de las columnas numericas
df_completo[['PJ_L','PG_L','PE_L','PP_L','GF_L','GC_L','DIF_L']] = df_completo[['PJ_L','PG_L','PE_L','PP_L','GF_L','GC_L','DIF_L']].astype('int64')

# Exportacion del DF Completo
df_completo.to_csv(r'C:\Users\Jorge\Documents\Portafolio V2\Python\Python_03_Rayados\Estadisticas_Rayados_Todo.csv',index = False)

df_completo
